Remove the Dockerfile
The node-helm pipeline is in charge of constructing the Dockerfile used
for building images. We have decided to use this approach because in the
pipeline, we can easily optimize the build by setting the node_modules
folder. This way, only the first build may take longer. Subsequent
builds will be faster because the modules would have been download by
previous builds. The build of other Node.js applications that use the
same dependencies will also benefit from this optimization.